Rename package(feature_favorite to feature.favorite)
