+++ app/src/main/java/com/hanbitkang/mypokemon/navigation/MpNavHost.kt
@@ -4,7 +4,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
-import com.hanbitkang.feature_favorite.navigation.favoriteGraph
+import com.hanbitkang.feature.favorite.navigation.favoriteGraph
 import com.hanbitkang.feature_pokemon.navigation.PokemonDestination
 import com.hanbitkang.feature_pokemon.navigation.pokemonGraph
