+++ feature-favorite/src/main/java/com/hanbitkang/feature/favorite/FavoriteScreen.kt
@@ -1,4 +1,4 @@
-package com.hanbitkang.feature_favorite
+package com.hanbitkang.feature.favorite
 
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
